@@ -22,75 +22,74 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.1",
-    "@angular/cdk": "^6.4.2",
-    "@angular/common": "^6.1.1",
-    "@angular/compiler": "^6.1.1",
-    "@angular/core": "^6.1.1",
-    "@angular/flex-layout": "^6.0.0-beta.18",
-    "@angular/forms": "^6.1.1",
-    "@angular/http": "^6.1.1",
-    "@angular/material": "^6.4.2",
-    "@angular/platform-browser": "^6.1.1",
-    "@angular/platform-browser-dynamic": "^6.1.1",
-    "@angular/router": "^6.1.1",
-    "@auth0/angular-jwt": "^2.0.0",
-    "@ngrx/effects": "^6.1.2",
-    "@ngrx/store": "^6.1.2",
-    "@ngrx/store-devtools": "^6.1.2",
-    "@swimlane/ngx-datatable": "^13.1.0",
-    "ajv": "^6.5.4",
+    "@angular/animations": "^7.2.15",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/core": "^7.2.15",
+    "@angular/forms": "^7.2.15",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/router": "^7.2.15",
+    "@auth0/angular-jwt": "^2.1.0",
+    "@ngrx/effects": "^7.4.0",
+    "@ngrx/store": "^7.4.0",
+    "@ngrx/store-devtools": "^7.4.0",
+    "@swimlane/ngx-datatable": "^14.0.0",
+    "ajv": "^6.10.0",
     "angular-password-strength-meter": "0.0.4",
     "core-js": "^2.5.7",
-    "css-element-queries": "^1.0.2",
-    "cytoscape": "^3.2.15",
-    "cytoscape-cola": "^2.2.3",
+    "css-element-queries": "^1.2.0",
+    "cytoscape": "^3.5.7",
+    "cytoscape-cola": "^2.3.0",
     "cytoscape-dagre": "^2.2.2",
     "cytoscape-spread": "^3.0.0",
-    "d3": "^5.5.0",
+    "d3": "^5.9.2",
     "elkjs": "^0.4.1",
     "hammerjs": "^2.0.8",
-    "html2canvas": "^1.0.0-alpha.12",
+    "html2canvas": "^1.0.0-rc.1",
     "immutable": "^3.8.2",
     "javascript-lp-solver": "^0.4.5",
-    "lodash": "^4.17.10",
+    "lodash": "^4.17.11",
     "material-design-icons": "^3.0.1",
-    "mxgraph": "^3.9.8",
+    "mxgraph": "^3.9.12",
     "normalize.css": "^8.0.1",
     "openlayers": "^4.6.5",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "zone.js": "^0.8.26",
+    "rxjs": "^6.5.2",
+    "tslib": "^1.9.3",
+    "zone.js": "^0.8.29",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.1.1",
-    "@types/d3": "^5.0.0",
-    "@types/hammerjs": "^2.0.35",
-    "@types/jasmine": "^2.8.8",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.9",
+    "@angular/compiler-cli": "^7.2.15",
+    "@types/d3": "^5.7.2",
+    "@types/hammerjs": "^2.0.36",
+    "@types/jasmine": "^3.3.12",
     "@types/node": "^10.5.6",
-    "@types/openlayers": "^4.6.10",
+    "@types/openlayers": "^4.6.17",
     "@types/zxcvbn": "^4.4.0",
-    "codelyzer": "^4.4.3",
-    "jasmine-core": "^3.1.0",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.4.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^2.0.5",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^1.2.0",
-    "protractor": "^5.4.0",
-    "rimraf": "^2.6.2",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "protractor": "^5.4.2",
+    "rimraf": "^2.6.3",
     "rxjs-tslint": "^0.1.7",
-    "rxjs-tslint-rules": "^4.20.0",
-    "stylelint": "^10.0.0",
+    "rxjs-tslint-rules": "^4.23.1",
+    "stylelint": "^10.0.1",
     "stylelint-config-standard": "^18.3.0",
     "ts-node": "^7.0.0",
-    "tslint": "^5.11.0",
+    "tslint": "^5.16.0",
     "tslint-config-standard": "^8.0.1",
-    "typescript": "^2.9.2"
+    "typescript": "^3.2.4"
   }
 }
